ABC algorithm for network reverse-engineering

ABC algorithm for network reverse-engineering

Usage

abc(
  data,
  clust_coeffs = c(0.33, 0.66, 1),
  tolerance = NA,
  number_hubs = NA,
  iterations = 10,
  number_networks = 1000,
  hub_probs = NA,
  neighbour_probs = NA,
  is_probs = 1
)

Arguments

data

: Any microarray data in the form of a matrix (rows are genes and columns are time points)

clust_coeffs

: one dimensional array of size clust_size of clustering coefficients (these clustering coefficient are tested in the ABc algorithm).

tolerance

: a positive real value based for the tolerance between the generated networks and the reference network

number_hubs

: number of hubs in the network

iterations

: number of times to repeat ABC algorithm

number_networks

: number of generated networks in each iteration of the ABC algorithm

hub_probs

: one-dimensional array of size number_genes for the each label to be in the role of a hub

neighbour_probs

: this is the matrix of neighbour probabilities of size number_nodes*number_nodes

is_probs

: this needs to be set either to one (if you specify hub_probs and neighbour_probs) or to zero (if neither probabilities are specified). Warning: you should specify both hub_probs and neighbour_probs if is_probs is one. If is_prob is zero these arrays should simply indicate an array of a specified size..

Examples

M<-matrix(rnorm(30),10,3)
result<-abc(data=M)
#> First run of abc to find tolerance
#> ===============================
#> Iteration=1
#> Accepted:1000
#> Probabilities of clustering coefficients:
#> 0.360000 0.345000 0.295000 
#> Tolerance value
#>       5% 
#> 4.754561 
#> ===============================
#> Beginning main run of abc
#> ===============================
#> Iteration=1
#> Accepted:51
#> Probabilities of clustering coefficients:
#> 0.333333 0.490196 0.176471 
#> ===============================
#> Iteration=2
#> Accepted:108
#> Probabilities of clustering coefficients:
#> 0.342593 0.490741 0.166667 
#> ===============================
#> Iteration=3
#> Accepted:130
#> Probabilities of clustering coefficients:
#> 0.346154 0.523077 0.130769 
#> ===============================
#> Iteration=4
#> Accepted:139
#> Probabilities of clustering coefficients:
#> 0.323741 0.553957 0.122302 
#> ===============================
#> Iteration=5
#> Accepted:155
#> Probabilities of clustering coefficients:
#> 0.335484 0.529032 0.135484 
#> ===============================
#> Iteration=6
#> Accepted:130
#> Probabilities of clustering coefficients:
#> 0.369231 0.461538 0.169231 
#> ===============================
#> Iteration=7
#> Accepted:143
#> Probabilities of clustering coefficients:
#> 0.475524 0.391608 0.132867 
#> ===============================
#> Iteration=8
#> Accepted:153
#> Probabilities of clustering coefficients:
#> 0.535948 0.339869 0.124183 
#> ===============================
#> Iteration=9
#> Accepted:229
#> Probabilities of clustering coefficients:
#> 0.628821 0.296943 0.074236 
#> ===============================
#> Iteration=10
#> Accepted:234
#> Probabilities of clustering coefficients:
#> 0.675214 0.273504 0.051282