ABC algorithm for network reverse-engineering

ABC algorithm for network reverse-engineering

abc(
  data,
  clust_coeffs = c(0.33, 0.66, 1),
  tolerance = NA,
  number_hubs = NA,
  iterations = 10,
  number_networks = 1000,
  hub_probs = NA,
  neighbour_probs = NA,
  is_probs = 1
)

Arguments

data

: Any microarray data in the form of a matrix (rows are genes and columns are time points)

clust_coeffs

: one dimensional array of size clust_size of clustering coefficients (these clustering coefficient are tested in the ABc algorithm).

tolerance

: a real value based for the tolerance between the generated networks and the reference network

number_hubs

: number of hubs in the network

iterations

: number of times to repeat ABC algorithm

number_networks

: number of generated networks in each iteration of the ABC algorithm

hub_probs

: one-dimensional array of size number_genes for the each label to be in the role of a hub

neighbour_probs

: this is the matrix of neighbour probabilities of size number_nodes*number_nodes

is_probs

: this needs to be set either to one (if you specify hub_probs and neighbour_probs) or to zero (if neither probabilities are specified). Warning: you should specify both hub_probs and neighbour_probs if is_probs is one. If is_prob is zero these arrays should simply indicate an array of a specified size..

Examples

M<-matrix(rnorm(30),10,3)
result<-abc(data=M)
#> First run of abc to find tolerance
#> ===============================
#> Iteration=1
#> Accepted:1000
#> Probabilities of clustering coefficients:
#> 0.346000 0.341000 0.313000 
#> Tolerance value
#>       5% 
#> 4.891706 
#> ===============================
#> Beginning main run of abc
#> ===============================
#> Iteration=1
#> Accepted:51
#> Probabilities of clustering coefficients:
#> 0.490196 0.313725 0.196078 
#> ===============================
#> Iteration=2
#> Accepted:169
#> Probabilities of clustering coefficients:
#> 0.573964 0.319527 0.106509 
#> ===============================
#> Iteration=3
#> Accepted:207
#> Probabilities of clustering coefficients:
#> 0.632850 0.275362 0.091787 
#> ===============================
#> Iteration=4
#> Accepted:271
#> Probabilities of clustering coefficients:
#> 0.708487 0.236162 0.055351 
#> ===============================
#> Iteration=5
#> Accepted:280
#> Probabilities of clustering coefficients:
#> 0.785714 0.178571 0.035714 
#> ===============================
#> Iteration=6
#> Accepted:309
#> Probabilities of clustering coefficients:
#> 0.883495 0.106796 0.009709 
#> ===============================
#> Iteration=7
#> Accepted:302
#> Probabilities of clustering coefficients:
#> 0.903974 0.092715 0.003311 
#> ===============================
#> Iteration=8
#> Accepted:347
#> Probabilities of clustering coefficients:
#> 0.919308 0.077810 0.002882 
#> ===============================
#> Iteration=9
#> Accepted:343
#> Probabilities of clustering coefficients:
#> 0.941691 0.055394 0.002915 
#> ===============================
#> Iteration=10
#> Accepted:324
#> Probabilities of clustering coefficients:
#> 0.947531 0.040123 0.012346