ABC algorithm for network reverse-engineering

ABC algorithm for network reverse-engineering

abc(
  data,
  clust_coeffs = c(0.33, 0.66, 1),
  tolerance = NA,
  number_hubs = NA,
  iterations = 10,
  number_networks = 1000,
  hub_probs = NA,
  neighbour_probs = NA,
  is_probs = 1
)

Arguments

data
: Any microarray data in the form of a matrix (rows are genes and columns are time points)

clust_coeffs
: one dimensional array of size clust_size of clustering coefficients (these clustering coefficient are tested in the ABc algorithm).

tolerance
: a real value based for the tolerance between the generated networks and the reference network

number_hubs
: number of hubs in the network

iterations
: number of times to repeat ABC algorithm

number_networks
: number of generated networks in each iteration of the ABC algorithm

hub_probs
: one-dimensional array of size number_genes for the each label to be in the role of a hub

neighbour_probs
: this is the matrix of neighbour probabilities of size number_nodes*number_nodes

is_probs
: this needs to be set either to one (if you specify hub_probs and neighbour_probs) or to zero (if neither probabilities are specified). Attention: you should specify both hub_probs and neighbour_probs if is_probs is one. If is_prob is zero these arrays should simply indicate an array of a specified size..

Examples

M<-matrix(rnorm(30),10,3)
result<-abc(data=M)

#> First run of abc to find tolerance
#> ===============================
#> Iteration=1
#> Accepted:1000
#> Probabilities of clustering coefficients:
#> 0.361000 0.330000 0.309000 
#> Tolerance value
#>       5% 
#> 4.874755 
#> ===============================
#> Beginning main run of abc
#> ===============================
#> Iteration=1
#> Accepted:49
#> Probabilities of clustering coefficients:
#> 0.326531 0.367347 0.306122 
#> ===============================
#> Iteration=2
#> Accepted:93
#> Probabilities of clustering coefficients:
#> 0.376344 0.408602 0.215054 
#> ===============================
#> Iteration=3
#> Accepted:104
#> Probabilities of clustering coefficients:
#> 0.442308 0.403846 0.153846 
#> ===============================
#> Iteration=4
#> Accepted:141
#> Probabilities of clustering coefficients:
#> 0.567376 0.347518 0.085106 
#> ===============================
#> Iteration=5
#> Accepted:145
#> Probabilities of clustering coefficients:
#> 0.579310 0.337931 0.082759 
#> ===============================
#> Iteration=6
#> Accepted:171
#> Probabilities of clustering coefficients:
#> 0.643275 0.350877 0.005848 
#> ===============================
#> Iteration=7
#> Accepted:182
#> Probabilities of clustering coefficients:
#> 0.670330 0.324176 0.005495 
#> ===============================
#> Iteration=8
#> Accepted:179
#> Probabilities of clustering coefficients:
#> 0.709497 0.290503 0.000000 
#> ===============================
#> Iteration=9
#> Accepted:178
#> Probabilities of clustering coefficients:
#> 0.741573 0.258427 0.000000 
#> ===============================
#> Iteration=10
#> Accepted:203
#> Probabilities of clustering coefficients:
#> 0.798030 0.201970 0.000000